requirements.txt
Flask
gunicorn
scikit-learn
pandas
numpy
pyyaml
pytest
emoji
